+++ lib/testclass.rb
@@ -1,4 +1,5 @@
 require_relative "zengin_bank_gem.rb"
+require_relative "bank.rb"
 
 class TestClass
   include Zengin
